+++ src/app/Editor/Panes/StructurePane/PrefixClasses.tsx
@@ -5,8 +5,9 @@ import Button from "@/app/shared/Button.js";
 import Group from "@/app/shared/Group.js";
 import { Input } from "@/app/shared/Input.js";
 import { performChange } from "@/lib/performChange.js";
+import type { IPrefixClassOptions } from "@/lib/transformers/index.js";
 import { prefixClasses } from "@/lib/transformers/index.js";
-import type { IPrefixClassOptions, PaneSubComponent } from "@/lib/types.js";
+import type { PaneSubComponent } from "@/lib/types.js";
 
 const PrefixClasses: PaneSubComponent = (props) => {
 	const [prefix, setPrefix] = useState("");
