+++ src/lib/overlay.ts
@@ -1,13 +1,9 @@
 import type React from "react";
 
+import type { IMovePointOptions } from "./transformers/index.js";
 import { movePoint } from "./transformers/index.js";
 
-import type {
-	ChangeOperation,
-	IMovePointOptions,
-	IPoint,
-	SVGSubElement,
-} from "@/lib/types.js";
+import type { ChangeOperation, IPoint, SVGSubElement } from "@/lib/types.js";
 
 /** The pixel width/height of the marker that shows on points. */
 const markerSize = 8;
