+++ src/lib/transformers/prefixClasses.ts
@@ -1,7 +1,6 @@
 import { renameClass } from "./renameClass.js";
 
 import { getClasses } from "@/lib/getClasses.js";
-import type { IPrefixClassOptions } from "@/lib/types.js";
 
 /**
  * Create a random prefix for the class
@@ -10,6 +9,10 @@ import type { IPrefixClassOptions } from "@/lib/types.js";
  */
 function createPrefix(): string {
 	return "svg-magic-" + Date.now().toString(36);
+}
+
+export interface IPrefixClassOptions {
+	prefix?: string;
 }
 
 /**
